+++ CHANGELOG.md
@@ -23,6 +23,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - gen_class.Mail.send_mailx:  Emails message out using mailx.
 
 ### Changed
+- gen_libs.chk_crt_file:  Replaced section of code on permission checks with call to perm_check.
 - gen_libs.chk_crt_file:  Added check to see if file is executable.
 - gen_class.Mail.send_mail:  Add a keyword argument option to allow the send_mailx method to be called.
 - gen_libs.print_dict:  Added ability to add the dictionary to an email instance message.